Исправления 3.11(В тёмной теме названия станций при выборе пункта отправления и назначения сливаются с фоном — они не читаемы)
