+++ TravelSchedule/View/RouteInputView.swift
@@ -6,29 +6,6 @@
 //
 import SwiftUI
 
-// Добавьте этот код в новый файл RouteInfo.swift или в начало RouteInputView.swift
-struct RouteInfo: Hashable {
-    let fromCity: City?
-    let fromStation: Station?
-    let toCity: City?
-    let toStation: Station?
-    
-    // Реализация Hashable для использования в NavigationStack
-    func hash(into hasher: inout Hasher) {
-        hasher.combine(fromCity?.name)
-        hasher.combine(fromStation?.name)
-        hasher.combine(toCity?.name)
-        hasher.combine(toStation?.name)
-    }
-    
-    static func == (lhs: RouteInfo, rhs: RouteInfo) -> Bool {
-        lhs.fromCity?.name == rhs.fromCity?.name &&
-        lhs.fromStation?.name == rhs.fromStation?.name &&
-        lhs.toCity?.name == rhs.toCity?.name &&
-        lhs.toStation?.name == rhs.toStation?.name
-    }
-}
-
 struct RouteInputView: View {
     @StateObject private var viewModel = RouteInputViewModel()
     @State private var navigationPath = NavigationPath()
